Consolidates BIG-IP-addressing changes into its own module for clarity.  The neutron changes will reflect similarly
